Crash page
This will capture any Exception that had not previously been caught. As a result this will prevent the normal "Unformtunately, APP has stopped." message. This screen is more useful because it shows why the app crashed and gives an easily allows someone to email me the issue (even without the market report button). It also provides some much needed humor :)